Renamed the modules of artifacts.
