+++ crates/bld_runner/src/pipeline/version2.rs
@@ -1,4 +1,4 @@
-use crate::artifacts::version2::Artifacts;
+use crate::artifacts::v2::Artifacts;
 use crate::external::version2::External;
 use crate::platform::version2::Platform;
 use crate::step::version2::BuildStep;
